+++ app/page.js
@@ -7,7 +7,7 @@ export default function Home() {
   return (
     <>
     <div className={styles.flexHeader}>
-      <p>Version: <span id="version-number" style={{color: '#01a0bb'}}>1.1.0-SFL</span></p>
+      <p>Version: <span id="version-number" style={{color: '#01a0bb'}}>1.1.0</span></p>
       <Button component="a" href="/db-view">
         View Data &#8594;
       </Button>
@@ -22,8 +22,6 @@ export default function Home() {
 
     <div className={styles.menuContainer}>
         <h1>Select Survey Mode</h1>
-        {/* <button class="btn menu-btn"><a href="./pit_survey/pit_form.html">Pit Survey</a></button>
-        <button class="btn menu-btn" disabled><a href="./match_survey/match_form.html">Match Survey</a></button> */}
         <div className={styles.buttonContainer}>
         <Button component="a" href="/pit-survey">
           Pit Survey